CHE-2552: Refactor the names of the packages
